refactor: 상품 가격(price) 타입을 Long으로 통일

+++ src/main/java/com/uhdyl/backend/product/dto/request/ProductCreateRequest.java
@@ -7,6 +7,6 @@ public record ProductCreateRequest(
         Category category,
         String breed,
         List<String> images,
-        int price,
+        Long price,
         String tone
 ){}
